move_files_function/requirements.txt
functions-framework
google-cloud-storage
Flask